Update MultivariateMCMC.py
Acceptance rate should accurately reflect the rate of acceptance following the burn-in period.

The in_burn_in attribute is set to True at the start of the burn-in period and back to False at the end. Then, in the step method, the acceptance count is only incremented if in_burn_in is False. This way, the acceptance rate only considers steps after the burn-in period, and you can easily check whether you're currently in the burn-in period by checking the value of self.in_burn_in.