Update MultivariateMCMC.py
Now considers the Jacobian adjustment in the target density which is required when a transformation is applied. In general, if you change variables in a density function, you need to adjust for the change of volume in the parameter space. The new density is the old density times the absolute value of the determinant of the Jacobian of the transformation.