Update MultivariateMCMC.py
***Change to sample method***
Store the old state before calling the step method. After the step method is called, check if the current state has changed (i.e., if the proposal was accepted). If the proposal was accepted and the state changed, store the new state in the samples array. If the proposal was not accepted and the state did not change, store the old state in the samples array. This way, only counts proposals as samples if they are accepted.